+++ CHANGELOG.md
@@ -87,3 +87,4 @@
 
 - ADDED addition and subtraction operators for Jalali
 - ADDED add method for Jalali
+- ADDED isValid, isLeapYear and monthLength methods to Gregorian